Adding collision event and modifying event listeners a bit
Now objects listen for events but not only does the game note what even
they're listening for but also the obejct itself so the game can find an
object to trigger its event if something interacts with it particularly.

Also, as an example of collisions, created door, which required a second
area. So added start of an overworld that's just a copy of world/dev for
the moment.